feat: JWT 기반 인증 시스템 구축
--- 본문 ---
- 기존 세션 기반 인증 방식에서 JWT 기반의 무상태(stateless) 인증 시스템으로 전환했습니다.
- 액세스 토큰 및 리프레시 토큰 발급, 검증, 인증 처리를 담당하는 `JwtTokenProvider`와 `JwtAuthenticationFilter`를 도입했습니다.
- 리프레시 토큰의 영속성을 위한 `jwt_refresh_tokens` 테이블을 추가했습니다.
- JWT 관련 오류 처리를 위한 `JwtErrorCode` 및 `JwtException`을 정의했습니다.
- `DiscodeitAuthorityUtils`를 `global.security.utils` 패키지로 이동하고, JWT 클레임에 사용될 권한 직렬화/역직렬화 기능을 추가했습니다.
- `DiscodeitUserDetails`에 `CredentialsContainer`를 구현하여 인증 정보 삭제 기능을 추가했습니다.
- `JwtProperties`의 토큰 만료 시간 단위를 `long`으로 변경하여 유연성을 높였습니다.
- `SecurityConfig`, `MethodSecurityConfig`, `QueryDslConfig` 등 보안 및 공통 설정 클래스의 의존성 주입 방식을 개선했습니다.

+++ src/main/java/com/sprint/mission/discodeit/global/security/jwt/JwtProperties.java
@@ -4,7 +4,7 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "discodeit.jwt")
 public record JwtProperties(String secretKey,
-                            int accessTokenExpiration,
-                            int refreshTokenExpiration) {
+                            long accessTokenExpiration,
+                            long refreshTokenExpiration) {
 
 }
